test(action-sheet): migrate translucent test
